fix heap of shit

+++ Il2CppInspector.Common/Reflection/Assembly.cs
@@ -28,7 +28,7 @@ namespace Il2CppInspector.Reflection {
         public string ShortName { get; }
 
         // Metadata token of the assembly
-        public int MetadataToken { get; }
+        public uint MetadataToken { get; }
 
         // Entry point method for the assembly
         public MethodInfo EntryPoint => throw new NotImplementedException();
@@ -48,7 +48,7 @@ namespace Il2CppInspector.Reflection {
             if (AssemblyDefinition.imageIndex != imageIndex)
                 throw new InvalidOperationException("Assembly/image index mismatch");
 
-            MetadataToken = (int) AssemblyDefinition.token;
+            MetadataToken =  AssemblyDefinition.token;
             Index = ImageDefinition.assemblyIndex;
             ShortName = Model.Package.Strings[ImageDefinition.nameIndex];
