+++ Il2CppInspector.Common/Reflection/PropertyInfo.cs
@@ -52,7 +52,7 @@ namespace Il2CppInspector.Reflection {
             base(declaringType) {
             Index = propIndex;
             Definition = pkg.Properties[propIndex];
-            MetadataToken = (int) Definition.token;
+            MetadataToken =  Definition.token;
             Name = pkg.Strings[Definition.nameIndex];
             rootDefinition = this;
